lambda/document-generator/requirements.txt
python-docx
reportlab
python-pptx